testing/web-platform/tests/workers/interfaces/WorkerUtils/navigator/001.html

<!--
for (var x in navigator) {
  postMessage(["Property", x]);
}
postMessage(["Done"]);
/*
-->
<!doctype html>
<title>navigator</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
async_test(function() {
  var expected = [
    'appCodeName',
    'appName',
    'appVersion',
    'platform',
    'product',
    'taintEnabled',
    'userAgent',
    'language',
    'languages',
    'onLine'
  ];
  var worker = new Worker('#');
  worker.onmessage = this.step_func(function(e) {
    switch (e.data[0]) {
    case "Property":
      var property = e.data[1];
      var index = expected.indexOf(property);
      assert_true(index >= 0, "Unexpected property " + format_value(property));
      expected.splice(index, 1);
      break;
    case "Done":
      assert_equals(expected.length, 0, "Missing properties: " + expected);
      this.done();
      break;
    default:
      assert_unreached("Unexpected message");
    }
  });
  worker.onerror = this.step_func(function(e) {
    assert_unreached('got error event');
  });
});
</script>
<!--
*/
//-->
